refactor(data_handling_pro): shorten module names structure_preprocessor->struct_preproc, annotation_preprocessor->annot_preproc
Renames the frontend modules and _backend subpackages to the short forms and
updates all mock/import path references in the pro preprocessor tests. Public
class names (StructurePreprocessor / AnnotationPreprocessor) are unchanged.
