+++ aaanalysis/data_handling/_embed_preproc.py
@@ -86,6 +86,7 @@ def _check_match_df_scales_df_stds(df_scales, df_stds):
 
 
 # II Main Functions
+# TODO THE
 class EmbeddingPreprocessor:
     """
     Utility data preprocessing class for protein-language-model (PLM) embeddings.
