+++ aaanalysis/data_handling_pro/__init__.py
@@ -1,5 +1,5 @@
-from ._structure_preprocessor import StructurePreprocessor
-from ._annotation_preprocessor import AnnotationPreprocessor
+from ._struct_preproc import StructurePreprocessor
+from ._annot_preproc import AnnotationPreprocessor
 
 __all__ = [
     "StructurePreprocessor",
