+++ aaanalysis/data_handling_pro/_backend/annot_preproc/_uniprot.py
