+++ aaanalysis/data_handling_pro/_backend/struct_preproc/align_dssp_full.py
