+++ tests/unit/data_handling_pro_tests/test_structure_preprocessor_encode_dssp.py
@@ -16,7 +16,7 @@ aa.options["verbose"] = False
 settings.register_profile("ci", deadline=2000)
 settings.load_profile("ci")
 
-MODULE = "aaanalysis.data_handling_pro._structure_preprocessor"
+MODULE = "aaanalysis.data_handling_pro._struct_preproc"
 RUNNER = f"{MODULE}.run_dssp_full_for_entry_"
